Remove lite_parse file
