+++ crates/nu-parser/src/lib.rs
@@ -3,7 +3,6 @@ mod errors;
 mod flatten;
 mod known_external;
 mod lex;
-mod lite_parse;
 mod parse_keywords;
 mod parser;
 mod type_check;
@@ -13,12 +12,12 @@ pub use errors::ParseError;
 pub use flatten::{flatten_block, flatten_expression, flatten_pipeline, FlatShape};
 pub use known_external::KnownExternal;
 pub use lex::{lex, Token, TokenContents};
-pub use lite_parse::{lite_parse, LiteBlock};
 pub use parse_keywords::*;
 
 pub use parser::{
-    is_math_expression_like, parse, parse_block, parse_duration_bytes, parse_expression,
-    parse_external_call, trim_quotes, trim_quotes_str, unescape_unquote_string, Import,
+    is_math_expression_like, lite_parse, parse, parse_block, parse_duration_bytes,
+    parse_expression, parse_external_call, trim_quotes, trim_quotes_str, unescape_unquote_string,
+    Import, LiteBlock,
 };
 
 #[cfg(feature = "plugin")]
