+++ crates/nu-parser/src/parser.rs
@@ -1,9 +1,7 @@
 use crate::{
-    lex, lite_parse,
-    lite_parse::LiteCommand,
-    parse_mut,
+    lex, parse_mut,
     type_check::{math_result_type, type_compatible},
-    LiteBlock, ParseError, Token, TokenContents,
+    ParseError, Token, TokenContents,
 };
 
 use nu_protocol::{
@@ -1906,9 +1904,6 @@ pub fn parse_full_cell_path(
             let source = working_set.get_span_contents(span);
 
             let (output, err) = lex(source, span.start, &[b'\n', b'\r'], &[], true);
-            error = error.or(err);
-
-            let (output, err) = lite_parse(&output);
             error = error.or(err);
 
             // Creating a Type scope to parse the new block. This will keep track of
@@ -3919,9 +3914,6 @@ pub fn parse_block_expression(
         _ => (None, 0),
     };
 
-    let (output, err) = lite_parse(&output[amt_to_skip..]);
-    error = error.or(err);
-
     // TODO: Finish this
     if let SyntaxShape::Closure(Some(v)) = shape {
         if let Some((sig, sig_span)) = &signature {
@@ -3955,8 +3947,13 @@ pub fn parse_block_expression(
         }
     }
 
-    let (mut output, err) =
-        parse_block(working_set, &output, false, expand_aliases_denylist, false);
+    let (mut output, err) = parse_block(
+        working_set,
+        &output[amt_to_skip..],
+        false,
+        expand_aliases_denylist,
+        false,
+    );
     error = error.or(err);
 
     if let Some(signature) = signature {
@@ -4088,9 +4085,6 @@ pub fn parse_closure_expression(
         _ => (None, 0),
     };
 
-    let (output, err) = lite_parse(&output[amt_to_skip..]);
-    error = error.or(err);
-
     // TODO: Finish this
     if let SyntaxShape::Closure(Some(v)) = shape {
         if let Some((sig, sig_span)) = &signature {
@@ -4124,8 +4118,13 @@ pub fn parse_closure_expression(
         }
     }
 
-    let (mut output, err) =
-        parse_block(working_set, &output, false, expand_aliases_denylist, false);
+    let (mut output, err) = parse_block(
+        working_set,
+        &output[amt_to_skip..],
+        false,
+        expand_aliases_denylist,
+        false,
+    );
     error = error.or(err);
 
     if let Some(signature) = signature {
@@ -5130,19 +5129,22 @@ pub fn parse_record(
 
 pub fn parse_block(
     working_set: &mut StateWorkingSet,
-    lite_block: &LiteBlock,
+    tokens: &[Token],
     scoped: bool,
     expand_aliases_denylist: &[usize],
     is_subexpression: bool,
 ) -> (Block, Option<ParseError>) {
+    let mut error = None;
+
+    let (lite_block, err) = lite_parse(tokens);
+    error = error.or(err);
+
     trace!("parsing block: {:?}", lite_block);
 
     if scoped {
         working_set.enter_scope();
     }
     working_set.type_scope.enter_scope();
-
-    let mut error = None;
 
     // Pre-declare any definition so that definitions
     // that share the same block can see each other
@@ -5618,6 +5620,188 @@ fn wrap_expr_with_collect(working_set: &mut StateWorkingSet, expr: &Expression) 
     }
 }
 
+#[derive(Debug)]
+pub struct LiteCommand {
+    pub comments: Vec<Span>,
+    pub parts: Vec<Span>,
+}
+
+impl Default for LiteCommand {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl LiteCommand {
+    pub fn new() -> Self {
+        Self {
+            comments: vec![],
+            parts: vec![],
+        }
+    }
+
+    pub fn push(&mut self, span: Span) {
+        self.parts.push(span);
+    }
+
+    pub fn is_empty(&self) -> bool {
+        self.parts.is_empty()
+    }
+}
+
+#[derive(Debug)]
+pub struct LitePipeline {
+    pub commands: Vec<LiteCommand>,
+}
+
+impl Default for LitePipeline {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl LitePipeline {
+    pub fn new() -> Self {
+        Self { commands: vec![] }
+    }
+
+    pub fn push(&mut self, command: LiteCommand) {
+        self.commands.push(command);
+    }
+
+    pub fn is_empty(&self) -> bool {
+        self.commands.is_empty()
+    }
+}
+
+#[derive(Debug)]
+pub struct LiteBlock {
+    pub block: Vec<LitePipeline>,
+}
+
+impl Default for LiteBlock {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl LiteBlock {
+    pub fn new() -> Self {
+        Self { block: vec![] }
+    }
+
+    pub fn push(&mut self, pipeline: LitePipeline) {
+        self.block.push(pipeline);
+    }
+
+    pub fn is_empty(&self) -> bool {
+        self.block.is_empty()
+    }
+}
+
+pub fn lite_parse(tokens: &[Token]) -> (LiteBlock, Option<ParseError>) {
+    let mut block = LiteBlock::new();
+    let mut curr_pipeline = LitePipeline::new();
+    let mut curr_command = LiteCommand::new();
+
+    let mut last_token = TokenContents::Eol;
+
+    let mut curr_comment: Option<Vec<Span>> = None;
+
+    for token in tokens.iter() {
+        match &token.contents {
+            TokenContents::Item => {
+                // If we have a comment, go ahead and attach it
+                if let Some(curr_comment) = curr_comment.take() {
+                    curr_command.comments = curr_comment;
+                }
+                curr_command.push(token.span);
+                last_token = TokenContents::Item;
+            }
+            TokenContents::Pipe => {
+                if !curr_command.is_empty() {
+                    curr_pipeline.push(curr_command);
+                    curr_command = LiteCommand::new();
+                }
+                last_token = TokenContents::Pipe;
+            }
+            TokenContents::Eol => {
+                if last_token != TokenContents::Pipe {
+                    if !curr_command.is_empty() {
+                        curr_pipeline.push(curr_command);
+
+                        curr_command = LiteCommand::new();
+                    }
+
+                    if !curr_pipeline.is_empty() {
+                        block.push(curr_pipeline);
+
+                        curr_pipeline = LitePipeline::new();
+                    }
+                }
+
+                if last_token == TokenContents::Eol {
+                    // Clear out the comment as we're entering a new comment
+                    curr_comment = None;
+                }
+
+                last_token = TokenContents::Eol;
+            }
+            TokenContents::Semicolon => {
+                if !curr_command.is_empty() {
+                    curr_pipeline.push(curr_command);
+
+                    curr_command = LiteCommand::new();
+                }
+
+                if !curr_pipeline.is_empty() {
+                    block.push(curr_pipeline);
+
+                    curr_pipeline = LitePipeline::new();
+                }
+
+                last_token = TokenContents::Semicolon;
+            }
+            TokenContents::Comment => {
+                // Comment is beside something
+                if last_token != TokenContents::Eol {
+                    curr_command.comments.push(token.span);
+                    curr_comment = None;
+                } else {
+                    // Comment precedes something
+                    if let Some(curr_comment) = &mut curr_comment {
+                        curr_comment.push(token.span);
+                    } else {
+                        curr_comment = Some(vec![token.span]);
+                    }
+                }
+
+                last_token = TokenContents::Comment;
+            }
+        }
+    }
+
+    if !curr_command.is_empty() {
+        curr_pipeline.push(curr_command);
+    }
+
+    if !curr_pipeline.is_empty() {
+        block.push(curr_pipeline);
+    }
+
+    if last_token == TokenContents::Pipe {
+        (
+            block,
+            Some(ParseError::UnexpectedEof(
+                "pipeline missing end".into(),
+                tokens[tokens.len() - 1].span,
+            )),
+        )
+    } else {
+        (block, None)
+    }
+}
+
 // Parses a vector of u8 to create an AST Block. If a file name is given, then
 // the name is stored in the working set. When parsing a source without a file
 // name, the source of bytes is stored as "source"
@@ -5642,9 +5826,6 @@ pub fn parse(
     working_set.add_file(name, contents);
 
     let (output, err) = lex(contents, span_offset, &[], &[], false);
-    error = error.or(err);
-
-    let (output, err) = lite_parse(&output);
     error = error.or(err);
 
     let (mut output, err) =
